[coverity] 1543039 Use of 32-bit time_t

+++ libfreerdp/core/input.h
@@ -38,7 +38,7 @@ typedef struct
 	rdpInputProxy* proxy;
 	wMessageQueue* queue;
 
-	UINT32 lastInputTimestamp;
+	UINT64 lastInputTimestamp;
 	UINT16 lastX;
 	UINT16 lastY;
 } rdp_input_internal;
